requirements.txt
alembic==1.9.4
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bandit==1.7.5
blinker==1.5
click==8.0.3
decorator==5.1.1
executing==0.8.3
Faker==12.0.1
flake8==4.0.1
Flask==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==1.1.2
iniconfig==1.1.1
ipython==8.10.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.2.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mdurl==0.1.2
mixer==7.2.2
packaging==21.3
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
pytest-env==0.6.2
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
rich==13.3.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.29
stack-data==0.2.0
stevedore==5.0.0
tomli==2.0.1
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
WTForms==3.0.1
